Finished the rough prototype.
I finished a prototype with no errors so far. This only supports Windows. It will track the amount of time you spent in 1 sitting. It tracks the amount of time you have spent in that day. It also tracks the total amount of time you have spent on the project.